requirements.txt
# ETL Pipeline Requirements
# Core ETL dependencies
pandas>=2.0.0
mysql-connector-python>=8.0.0
requests>=2.28.0
python-dotenv>=1.0.0

# GUI dependencies
PySide6>=6.0.0

# Optional development dependencies
# pytest>=7.0.0  # For extended testing
# black>=22.0.0  # Code formatting
# flake8>=5.0.0  # Code linting